requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
sqlalchemy>=2.0.29
psycopg2-binary>=2.9.12
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2
python-multipart>=0.0.9
pydantic[email]>=2.6.4
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
pytest>=8.1.1
httpx>=0.27.0
